................
  r8631 | dhustace | 2008-03-14 18:18:37 -0400 (Fri, 14 Mar 2008) | 1 line
  
  Fixing bug 2173
................
  r8633 | rangerrick | 2008-03-15 10:22:46 -0400 (Sat, 15 Mar 2008) | 9 lines
  
  Merged revisions 8632 via svnmerge from 
  https://opennms.svn.sourceforge.net/svnroot/opennms/opennms/trunk
  
  ........
    r8632 | mhuot | 2008-03-15 10:09:25 -0400 (Sat, 15 Mar 2008) | 1 line
    
    Fixes bug 2345, review by RangerRick(Thanks for the help!)
  ........
................
  r8635 | tarus | 2008-03-16 12:48:02 -0400 (Sun, 16 Mar 2008) | 2 lines
  
  Added Alvarion device support - bug 2332
................
  r8636 | djgregor | 2008-03-16 15:23:23 -0400 (Sun, 16 Mar 2008) | 1 line
  
  Use higher precision data types when doing conversions to long.  This makes us do pretty much the same thing in SNMP4J and JoeSNMP.
................
  r8637 | djgregor | 2008-03-16 16:03:37 -0400 (Sun, 16 Mar 2008) | 4 lines
  
  - Rename SnmpMonitorTest to SnmpMonitorStrategyTest because that is what is really being tested.
  - Add opennms-snmp-test test-jar to opennms-services pom.xml.
  - Use TestSnmpValue value objects in SnmpMonitorStrategyTest instead of objects specific to an SNMP implementation.
  - Add Integer32/Gauge32/Counter32/Counter64  value objects to TestSnmpValue.
................
  r8638 | djgregor | 2008-03-16 21:23:47 -0400 (Sun, 16 Mar 2008) | 1 line
  
  make opennms-snmp-test produce a test-jar so that opennms-services can consume it
................
  r8639 | djgregor | 2008-03-16 21:27:41 -0400 (Sun, 16 Mar 2008) | 1 line
  
  oops... add missing class to the last commit which had too much stuff
................
  r8640 | djgregor | 2008-03-16 21:29:46 -0400 (Sun, 16 Mar 2008) | 1 line
  
  Implement the MOAccess interface and mark these objects as unreadable, well, because they are.
................
  r8641 | djgregor | 2008-03-16 22:25:52 -0400 (Sun, 16 Mar 2008) | 1 line
  
  - Extend MockSnmpAgentTestCase.
................
  r8642 | djgregor | 2008-03-16 23:30:55 -0400 (Sun, 16 Mar 2008) | 14 lines
  
  Merged revisions 8579-8580 via svnmerge from 
  https://opennms.svn.sourceforge.net/svnroot/opennms/opennms/trunk
  
  ........
    r8579 | djgregor | 2008-03-05 23:11:11 -0500 (Wed, 05 Mar 2008) | 2 lines
    
    - Avoid catching a database exception in checkIndexUniqueness, do pre-query checks instead to make sure all columns exist.
    - Add a test for this case.
  ........
    r8580 | djgregor | 2008-03-05 23:23:32 -0500 (Wed, 05 Mar 2008) | 1 line
    
    Include the bug number in the test testCheckIndexUniquenessWithTableButMissingColumnBug2325.
  ........
................
  r8644 | brozow | 2008-03-17 11:35:56 -0400 (Mon, 17 Mar 2008) | 1 line
  
  bug #1871 fix misspelling of property name that prevents exitting
................
  r8646 | brozow | 2008-03-17 16:14:53 -0400 (Mon, 17 Mar 2008) | 1 line
  
  Bug #1296 make sure we don't send resolved notifications if notifications are globally turned off
................
  r8648 | brozow | 2008-03-17 22:40:01 -0400 (Mon, 17 Mar 2008) | 1 line
  
  Bug #2218 Add error page for when No applications exist
................
  r8650 | jgehlbach | 2008-03-18 00:38:20 -0400 (Tue, 18 Mar 2008) | 1 line
  
  Fix bug 2170, "ucd-sysstat uses deprecated OIDs in datacollection-config.xml".  Added new collection objects SysRawContext and SysRawInterrupts and corresponding prefab resource graphs.  Left old SysContext and SysInterrupts in place but removed their resource graphs from the default report list.
................
  r8651 | jgehlbach | 2008-03-18 00:47:55 -0400 (Tue, 18 Mar 2008) | 1 line
  
  Update release notes for resolution of bug 2170
................
  r8652 | jgehlbach | 2008-03-18 01:19:56 -0400 (Tue, 18 Mar 2008) | 1 line
  
  Fix bug 2252, "Acknowledge entire search for Alarms fails in 1.3.11".  Fix acknowledge_by_filter_form action attribute and actionCode input so that (un)acknowledging the entire search works again.
................
  r8653 | jgehlbach | 2008-03-18 01:48:44 -0400 (Tue, 18 Mar 2008) | 3 lines
  
  Fix bug 2287, "Applications can no longer be edited in the WebUI".
  
  The improved bounds-checking introduced in r7971 necessitates having a value for the "edit" URL parameter, so I went through and set the value to "edit".
................

+++ opennms-snmp/opennms-snmp-snmp4j/src/main/java/org/opennms/netmgt/snmp/snmp4j/Snmp4JValue.java
@@ -198,12 +198,12 @@ class Snmp4JValue implements SnmpValue {
         case SMIConstants.SYNTAX_OCTET_STRING:
             return (convertStringToLong());
         default:
-            return Integer.parseInt(m_value.toString());
+            return Long.parseLong(m_value.toString());
         }
     }
 
     private long convertStringToLong() {
-        return Float.valueOf(m_value.toString()).longValue();
+        return Double.valueOf(m_value.toString()).longValue();
     }
 
     public String toDisplayString() {
